third_party/blink/web_tests/xr/xrSession_environmentBlendMode.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/testdriver.js"></script>
<script src="../resources/testdriver-vendor.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="../external/wpt/webxr/resources/webxr_test_constants.js"></script>
<script src="../xr/resources/xr-test-utils.js"></script>
<canvas />

<script>

let immersiveTestName = "environmentBlendMode is correct for a VR device in immersive";
let nonImmersiveTestName = "environmentBlendMode is correct for a VR device in non-immersive";

let fakeDeviceInitParams = TRACKED_IMMERSIVE_DEVICE;

let testFunction = function(session, fakeDeviceController, t) {
  t.step(() => {
    assert_equals(session.environmentBlendMode, 'opaque');
  });
};

xr_session_promise_test(
  immersiveTestName, testFunction, fakeDeviceInitParams, 'immersive-vr');

xr_session_promise_test(
  nonImmersiveTestName, testFunction, fakeDeviceInitParams, 'inline');

// TODO(https://crbug.com/828321): Enable once session options for AR are in place.
/*xr_session_promise_test("environmentBlendMode is correct for an AR device",
 (session, fakeDeviceController, t) => {
  t.step(() => {
    assert_equals(session.environmentBlendMode, 'alpha-blend');
  });
}, fakeDevices["FakeARPhone"], [
    { ar: true, outputContext: getOutputContext() }
]);*/

</script>
